Fix some more things to make the build pass
