+++ packages/components/src/ui/utils/types.ts
@@ -1,16 +1,3 @@
-/**
- * External dependencies
- */
-import type { As } from 'reakit-utils/types';
-import type { ViewOwnProps } from '@wp-g2/create-styles';
-
-export type Options< T extends As, P extends ViewOwnProps< {}, T > > = {
-	as: T;
-	name: string;
-	useHook: ( props: P ) => any;
-	memo?: boolean;
-};
-
 export type ResponsiveCSSValue< T > = Array< T | undefined > | T;
 
 export type PopperPlacement =
